Moves SHA1 hash engine to common.

+++ common/sha.cpp
@@ -83,7 +83,7 @@ void SHAEngine::Process_Partial(void const*& data, long& length)
     **	Attach as many bytes as possible from the source data into
     **	the staging buffer.
     */
-    int add_count = std::min((int)length, SRC_BLOCK_SIZE - PartialCount);
+    int add_count = std::min((int)length, (int)SRC_BLOCK_SIZE - PartialCount);
     memcpy(&Partial[PartialCount], data, add_count);
     data = ((char const*&)data) + add_count;
     PartialCount += add_count;
